frontend/requirements.txt
streamlit
python-dotenv
langchain_community
langchain